Fix scheduled_time handling in ahi_hsd reader
Added min and sec information to scheduled_time of ahi_hsd reader.
The scheduled_time had an accuracy of 10 min, but the Japan region and Region 3 are observed at a frequency of 2.5 min.
Regions 4 and 5 of non-public data are observed at a frequency of 30 sec.
With this fix, scheduled_time is calculated correctly for those observations.